requirements.txt
google-genai
sentence-transformers
python-dotenv
ipython
pymupdf4llm
tqdm
